# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delfosti",
        "description": "Products Search Api",
        "contact": {
            "email": "marcoans@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://delfosti.com/wp-content/uploads/2022/06/Logo-secundario.png"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/products/search": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of searched products",
                "description": "Returns list of products",
                "operationId": "productsSearch",
                "parameters": [
                    {
                        "name": "term",
                        "in": "query",
                        "description": "Search term",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "blue"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "API Endpoints for Products"
        }
    ]
}